--- alembic/versions/0f1c29aacf5d_add_target_count.py
@@ -1,30 +1,0 @@
-"""add target count
-
-Revision ID: 0f1c29aacf5d
-Revises: 799ff9e1cef0
-Create Date: 2021-03-23 02:12:07.719428
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "0f1c29aacf5d"
-down_revision = "799ff9e1cef0"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("contract", sa.Column("deposit_lower", sa.Integer(), nullable=True))
-    op.add_column("contract", sa.Column("deposit_upper", sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("contract", "deposit_upper")
-    op.drop_column("contract", "deposit_lower")
-    # ### end Alembic commands ###
