--- alembic/versions/d56e3cf310de_init_base_model_entity.py
@@ -1,131 +1,0 @@
-"""Init base model entity
-
-Revision ID: d56e3cf310de
-Revises: 
-    Add Province table
-    Add Building table
-    Add Contract table
-
-Create Date: 2021-03-19 02:21:01.644521
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "d56e3cf310de"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-en_building_category = sa.Enum(
-    "APARTMENT",
-    "OFFICETEL",
-    name="en_building_category",
-)
-
-en_contract_category = sa.Enum(
-    "DEAL",
-    "TNANT",
-    "RENT",
-    name="en_contract_category",
-)
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "province",
-        sa.Column("created_datetime", sa.DateTime(), nullable=False),
-        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("sido", sa.String(length=32), nullable=True),
-        sa.Column("gugun", sa.String(length=32), nullable=True),
-        sa.Column("eupmyeondong", sa.String(length=32), nullable=True),
-        sa.Column("dongli", sa.String(length=32), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_province_created_datetime"),
-        "province",
-        ["created_datetime"],
-        unique=False,
-    )
-
-    op.create_table(
-        "building",
-        sa.Column("created_datetime", sa.DateTime(), nullable=False),
-        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("province_id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(length=256), nullable=False),
-        sa.Column("category", en_building_category, nullable=False),
-        sa.Column("dong", sa.String(length=64), nullable=True),
-        sa.Column("ho", sa.String(length=64), nullable=True),
-        sa.Column("floor", sa.Integer(), nullable=True),
-        sa.Column("max_floor", sa.Integer(), nullable=True),
-        sa.Column("exclusive_area", sa.Float(asdecimal=True), nullable=False),
-        sa.Column("shared_area", sa.Float(asdecimal=True), nullable=True),
-        sa.Column("window_side", sa.String(length=64), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["province_id"], ["province.id"], name="province_fkey", ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_building_created_datetime"),
-        "building",
-        ["created_datetime"],
-        unique=False,
-    )
-    op.create_index(op.f("ix_building_name"), "building", ["name"], unique=False)
-    op.create_index(
-        op.f("ix_building_province_id"), "building", ["province_id"], unique=False
-    )
-
-    op.create_table(
-        "contract",
-        sa.Column("created_datetime", sa.DateTime(), nullable=False),
-        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("building_id", sa.Integer(), nullable=False),
-        sa.Column("category", en_contract_category, nullable=False),
-        sa.Column("deal_upper", sa.Integer(), nullable=True),
-        sa.Column("deal_lower", sa.Integer(), nullable=True),
-        sa.Column("tnant_upper", sa.Integer(), nullable=True),
-        sa.Column("tnant_lower", sa.Integer(), nullable=True),
-        sa.Column("rent_upper", sa.Integer(), nullable=True),
-        sa.Column("rent_lower", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["building_id"], ["building.id"], name="building_fkey", ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_contract_building_id"), "contract", ["building_id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_contract_created_datetime"),
-        "contract",
-        ["created_datetime"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_contract_created_datetime"), table_name="contract")
-    op.drop_index(op.f("ix_contract_building_id"), table_name="contract")
-    op.drop_table("contract")
-    op.drop_index(op.f("ix_building_province_id"), table_name="building")
-    op.drop_index(op.f("ix_building_name"), table_name="building")
-    op.drop_index(op.f("ix_building_created_datetime"), table_name="building")
-    op.drop_table("building")
-    op.drop_index(op.f("ix_province_created_datetime"), table_name="province")
-    op.drop_table("province")
-
-    en_building_category.drop(op.get_bind())
-    en_contract_category.drop(op.get_bind())
-    # ### end Alembic commands ###
