+++ alembic/versions/f4e50ed3f1a4_added_initial_table.py
@@ -1,0 +1,163 @@
+"""Added initial table
+
+Revision ID: f4e50ed3f1a4
+Revises: 
+Create Date: 2021-03-28 18:11:48.879209
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = "f4e50ed3f1a4"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+en_building_category = sa.Enum(
+    "APARTMENT",
+    "OFFICETEL",
+    name="en_building_category",
+)
+
+en_contract_category = sa.Enum(
+    "DEAL",
+    "TNANT",
+    "RENT",
+    name="en_contract_category",
+)
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "building",
+        sa.Column("created_datetime", sa.DateTime(), nullable=False),
+        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("name", sa.String(length=256), nullable=False),
+        sa.Column("category", en_building_category, nullable=False),
+        sa.Column("built_year", sa.Integer(), nullable=True),
+        sa.Column("total_dong", sa.Integer(), nullable=True),
+        sa.Column("total_household", sa.Integer(), nullable=True),
+        sa.Column("land_address", sa.String(length=128), nullable=True),
+        sa.Column("road_address", sa.String(length=128), nullable=True),
+        sa.Column("deal_count", sa.Integer(), nullable=True),
+        sa.Column("tnant_count", sa.Integer(), nullable=True),
+        sa.Column("rent_count", sa.Integer(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_building_created_datetime"),
+        "building",
+        ["created_datetime"],
+        unique=False,
+    )
+    op.create_index(op.f("ix_building_name"), "building", ["name"], unique=False)
+    op.create_table(
+        "areatype",
+        sa.Column("created_datetime", sa.DateTime(), nullable=False),
+        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("area_name", sa.String(length=16), nullable=False),
+        sa.Column("shared_area", sa.Float(), nullable=False),
+        sa.Column("exclusive_area", sa.Float(), nullable=False),
+        sa.Column("building_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["building_id"], ["building.id"], name="building_fkey", ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_areatype_building_id"), "areatype", ["building_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_areatype_created_datetime"),
+        "areatype",
+        ["created_datetime"],
+        unique=False,
+    )
+    op.create_table(
+        "household",
+        sa.Column("created_datetime", sa.DateTime(), nullable=False),
+        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("dong", sa.Integer(), nullable=True),
+        sa.Column("floor", sa.Integer(), nullable=True),
+        sa.Column("ho", sa.String(length=32), nullable=True),
+        sa.Column("window_side", sa.String(length=32), nullable=True),
+        sa.Column("building_id", sa.Integer(), nullable=False),
+        sa.Column("areatype_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["areatype_id"], ["areatype.id"], name="areatype_fkey", ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(
+            ["building_id"], ["building.id"], name="building_fkey", ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_household_areatype_id"), "household", ["areatype_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_household_building_id"), "household", ["building_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_household_created_datetime"),
+        "household",
+        ["created_datetime"],
+        unique=False,
+    )
+    op.create_table(
+        "contract",
+        sa.Column("created_datetime", sa.DateTime(), nullable=False),
+        sa.Column("modified_datetime", sa.DateTime(), nullable=True),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("category", en_contract_category, nullable=False),
+        sa.Column("deal_upper", sa.Integer(), nullable=True),
+        sa.Column("deal_lower", sa.Integer(), nullable=True),
+        sa.Column("deposit_upper", sa.Integer(), nullable=True),
+        sa.Column("deposit_lower", sa.Integer(), nullable=True),
+        sa.Column("rent_upper", sa.Integer(), nullable=True),
+        sa.Column("rent_lower", sa.Integer(), nullable=True),
+        sa.Column("household_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["household_id"],
+            ["household.id"],
+            name="household_fkey",
+            ondelete="CASCADE",
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_contract_created_datetime"),
+        "contract",
+        ["created_datetime"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_contract_household_id"), "contract", ["household_id"], unique=False
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f("ix_contract_household_id"), table_name="contract")
+    op.drop_index(op.f("ix_contract_created_datetime"), table_name="contract")
+    op.drop_table("contract")
+    op.drop_index(op.f("ix_household_created_datetime"), table_name="household")
+    op.drop_index(op.f("ix_household_building_id"), table_name="household")
+    op.drop_index(op.f("ix_household_areatype_id"), table_name="household")
+    op.drop_table("household")
+    op.drop_index(op.f("ix_areatype_created_datetime"), table_name="areatype")
+    op.drop_index(op.f("ix_areatype_building_id"), table_name="areatype")
+    op.drop_table("areatype")
+    op.drop_index(op.f("ix_building_name"), table_name="building")
+    op.drop_index(op.f("ix_building_created_datetime"), table_name="building")
+    op.drop_table("building")
+
+    en_building_category.drop(op.get_bind())
+    en_contract_category.drop(op.get_bind())
+    # ### end Alembic commands ###
